@@ -4,6 +4,11 @@
 // not use this file except in compliance with the License. You may obtain
 // a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
 
+// fuzz_test.go is used by the "oss-fuzz" integration. Use caution when
+// modifying this file because it may break that integration.
+//
+// See https://github.com/google/oss-fuzz/tree/master/projects/mongo-go-driver
+
 package bson
 
 import (
@@ -29,11 +34,11 @@ func FuzzDecode(f *testing.F) {
 
 			encoded, err := Marshal(i)
 			if err != nil {
-				t.Fatal("failed to marshal", err)
+				t.Fatalf("failed to marshal: %v", err)
 			}
 
 			if err := Unmarshal(encoded, i); err != nil {
-				t.Fatal("failed to unmarshal", err)
+				t.Fatalf("failed to unmarshal: %v", err)
 			}
 		}
 	})
