@@ -8,57 +8,60 @@ package unified
 
 import (
 	"context"
-	"path"
+	"os"
 	"testing"
+
+	"go.mongodb.org/mongo-driver/v2/internal/spectest"
 )
 
 var (
 	passDirectories = []string{
-		"unified-test-format/valid-pass",
-		"versioned-api",
-		"crud/unified",
-		"change-streams",
-		"transactions/unified",
-		"load-balancers",
-		"collection-management",
-		"command-monitoring",
-		"command-monitoring/logging",
-		"connection-monitoring-and-pooling/logging",
-		"sessions",
-		"retryable-reads/unified",
-		"retryable-writes/unified",
-		"client-side-encryption/unified",
-		"client-side-operations-timeout",
-		"gridfs",
-		"server-selection/logging",
-		"server-discovery-and-monitoring/unified",
-		"run-command",
-		"index-management",
+		"unified-test-format/tests/valid-pass",
+		"transactions-convenient-api/tests/unified",
+		"versioned-api/tests",
+		"crud/tests/unified",
+		"change-streams/tests/unified",
+		"load-balancers/tests",
+		"collection-management/tests",
+		"command-logging-and-monitoring/tests/monitoring",
+		"command-logging-and-monitoring/tests/logging",
+		"connection-monitoring-and-pooling/tests/logging",
+		"sessions/tests",
+		"retryable-reads/tests/unified",
+		"retryable-writes/tests/unified",
+		"client-side-encryption/tests/unified",
+		"client-side-operations-timeout/tests",
+		"gridfs/tests",
+		"server-selection/tests/logging",
+		"server-discovery-and-monitoring/tests/unified",
+		"run-command/tests/unified",
+		"index-management/tests",
+		"atlas-data-lake-testing/tests/unified",
 	}
 	failDirectories = []string{
-		"unified-test-format/valid-fail",
+		"unified-test-format/tests/valid-fail",
 	}
-)
-
-const (
-	dataDirectory = "../../../testdata"
 )
 
 func TestUnifiedSpec(t *testing.T) {
 	// Ensure the cluster is in a clean state before test execution begins.
-	if err := terminateOpenSessions(context.Background()); err != nil {
-		t.Fatalf("error terminating open transactions: %v", err)
+	// Don't run for Data Lake tests because it doesn't support
+	// "killAllSessions".
+	if os.Getenv("ATLAS_DATA_LAKE_INTEGRATION_TEST") != "true" {
+		if err := terminateOpenSessions(context.Background()); err != nil {
+			t.Fatalf("error terminating open transactions: %v", err)
+		}
 	}
 
 	for _, testDir := range passDirectories {
 		t.Run(testDir, func(t *testing.T) {
-			runTestDirectory(t, path.Join(dataDirectory, testDir), false)
+			runTestDirectory(t, spectest.Path(testDir), false)
 		})
 	}
 
 	for _, testDir := range failDirectories {
 		t.Run(testDir, func(t *testing.T) {
-			runTestDirectory(t, path.Join(dataDirectory, testDir), true)
+			runTestDirectory(t, spectest.Path(testDir), true)
 		})
 	}
 }
