@@ -70,12 +70,13 @@ func (dom *DeleteOneModel) SetCollation(collation *options.Collation) *DeleteOne
 	return dom
 }
 
-// SetHint specifies the index to use for the operation. This should either be the index name as a string or the index
-// specification as a document. This option is only valid for MongoDB versions >= 4.4. Server versions >= 3.4 will
-// return an error if this option is specified. For server versions < 3.4, the driver will return a client-side error if
-// this option is specified. The driver will return an error if this option is specified during an unacknowledged write
-// operation. The driver will return an error if the hint parameter is a multi-key map. The default value is nil, which
-// means that no hint will be sent.
+// SetHint specifies the index to use for the operation. This should either be
+// the index name as a string or the index specification as a document. This
+// option is only valid for MongoDB versions >= 4.4. Server versions < 4.4 will
+// return an error if this option is specified. The driver will return an error
+// if this option is specified during an unacknowledged write operation. The
+// driver will return an error if the hint parameter is a multi-key map. The
+// default value is nil, which means that no hint will be sent.
 func (dom *DeleteOneModel) SetHint(hint interface{}) *DeleteOneModel {
 	dom.Hint = hint
 	return dom
@@ -111,12 +112,13 @@ func (dmm *DeleteManyModel) SetCollation(collation *options.Collation) *DeleteMa
 	return dmm
 }
 
-// SetHint specifies the index to use for the operation. This should either be the index name as a string or the index
-// specification as a document. This option is only valid for MongoDB versions >= 4.4. Server versions >= 3.4 will
-// return an error if this option is specified. For server versions < 3.4, the driver will return a client-side error if
-// this option is specified. The driver will return an error if this option is specified during an unacknowledged write
-// operation. The driver will return an error if the hint parameter is a multi-key map. The default value is nil, which
-// means that no hint will be sent.
+// SetHint specifies the index to use for the operation. This should either be
+// the index name as a string or the index specification as a document. This
+// option is only valid for MongoDB versions >= 4.4. Server versions < 4.4 will
+// return an error if this option is specified. The driver will return an error
+// if this option is specified during an unacknowledged write operation. The
+// driver will return an error if the hint parameter is a multi-key map. The
+// default value is nil, which means that no hint will be sent.
 func (dmm *DeleteManyModel) SetHint(hint interface{}) *DeleteManyModel {
 	dmm.Hint = hint
 	return dmm
@@ -141,12 +143,13 @@ func NewReplaceOneModel() *ReplaceOneModel {
 	return &ReplaceOneModel{}
 }
 
-// SetHint specifies the index to use for the operation. This should either be the index name as a string or the index
-// specification as a document. This option is only valid for MongoDB versions >= 4.2. Server versions >= 3.4 will
-// return an error if this option is specified. For server versions < 3.4, the driver will return a client-side error if
-// this option is specified. The driver will return an error if this option is specified during an unacknowledged write
-// operation. The driver will return an error if the hint parameter is a multi-key map. The default value is nil, which
-// means that no hint will be sent.
+// SetHint specifies the index to use for the operation. This should either be
+// the index name as a string or the index specification as a document. This
+// option is only valid for MongoDB versions >= 4.2. Server versions < 4.2 will
+// return an error if this option is specified. The driver will return an error
+// if this option is specified during an unacknowledged write operation. The
+// driver will return an error if the hint parameter is a multi-key map. The
+// default value is nil, which means that no hint will be sent.
 func (rom *ReplaceOneModel) SetHint(hint interface{}) *ReplaceOneModel {
 	rom.Hint = hint
 	return rom
@@ -211,12 +214,13 @@ func NewUpdateOneModel() *UpdateOneModel {
 	return &UpdateOneModel{}
 }
 
-// SetHint specifies the index to use for the operation. This should either be the index name as a string or the index
-// specification as a document. This option is only valid for MongoDB versions >= 4.2. Server versions >= 3.4 will
-// return an error if this option is specified. For server versions < 3.4, the driver will return a client-side error if
-// this option is specified. The driver will return an error if this option is specified during an unacknowledged write
-// operation. The driver will return an error if the hint parameter is a multi-key map. The default value is nil, which
-// means that no hint will be sent.
+// SetHint specifies the index to use for the operation. This should either be
+// the index name as a string or the index specification as a document. This
+// option is only valid for MongoDB versions >= 4.2. Server versions < 4.2 will
+// return an error if this option is specified. The driver will return an error
+// if this option is specified during an unacknowledged write operation. The
+// driver will return an error if the hint parameter is a multi-key map. The
+// default value is nil, which means that no hint will be sent.
 func (uom *UpdateOneModel) SetHint(hint interface{}) *UpdateOneModel {
 	uom.Hint = hint
 	return uom
@@ -287,12 +291,13 @@ func NewUpdateManyModel() *UpdateManyModel {
 	return &UpdateManyModel{}
 }
 
-// SetHint specifies the index to use for the operation. This should either be the index name as a string or the index
-// specification as a document. This option is only valid for MongoDB versions >= 4.2. Server versions >= 3.4 will
-// return an error if this option is specified. For server versions < 3.4, the driver will return a client-side error if
-// this option is specified. The driver will return an error if this option is specified during an unacknowledged write
-// operation. The driver will return an error if the hint parameter is a multi-key map. The default value is nil, which
-// means that no hint will be sent.
+// SetHint specifies the index to use for the operation. This should either be
+// the index name as a string or the index specification as a document. This
+// option is only valid for MongoDB versions >= 4.2. Server versions < 4.2 will
+// return an error if this option is specified. The driver will return an error
+// if this option is specified during an unacknowledged write operation. The
+// driver will return an error if the hint parameter is a multi-key map. The
+// default value is nil, which means that no hint will be sent.
 func (umm *UpdateManyModel) SetHint(hint interface{}) *UpdateManyModel {
 	umm.Hint = hint
 	return umm
