@@ -141,16 +141,16 @@ func TestDatabase(t *testing.T) {
 		db := setupDb("foo")
 
 		err := db.RunCommand(bgCtx, nil).Err()
-		assert.Equal(t, ErrNilDocument, err, "expected error %v, got %v", ErrNilDocument, err)
+		assert.True(t, errors.Is(err, ErrNilDocument), "expected error %v, got %v", ErrNilDocument, err)
 
 		_, err = db.Watch(context.Background(), nil)
 		watchErr := errors.New("can only marshal slices and arrays into aggregation pipelines, but got invalid")
 		assert.Equal(t, watchErr, err, "expected error %v, got %v", watchErr, err)
 
 		_, err = db.ListCollections(context.Background(), nil)
-		assert.Equal(t, ErrNilDocument, err, "expected error %v, got %v", ErrNilDocument, err)
+		assert.True(t, errors.Is(err, ErrNilDocument), "expected error %v, got %v", ErrNilDocument, err)
 
 		_, err = db.ListCollectionNames(context.Background(), nil)
-		assert.Equal(t, ErrNilDocument, err, "expected error %v, got %v", ErrNilDocument, err)
+		assert.True(t, errors.Is(err, ErrNilDocument), "expected error %v, got %v", ErrNilDocument, err)
 	})
 }
