@@ -596,24 +596,45 @@ func (c *ClientOptions) Validate() error {
 			return fmt.Errorf("cannot set both OIDCMachineCallback and OIDCHumanCallback, only one may be specified")
 		}
 		if c.Auth.OIDCHumanCallback == nil && c.Auth.AuthMechanismProperties[auth.AllowedHostsProp] != "" {
-			return fmt.Errorf("Cannot specify ALLOWED_HOSTS without an OIDCHumanCallback")
+			return fmt.Errorf("cannot specify ALLOWED_HOSTS without an OIDCHumanCallback")
 		}
+		if c.Auth.OIDCMachineCallback == nil && c.Auth.OIDCHumanCallback == nil && c.Auth.AuthMechanismProperties[auth.EnvironmentProp] == "" {
+			return errors.New("must specify at least one of OIDCMachineCallback, OIDCHumanCallback, or ENVIRONMENT authMechanismProperty")
+		}
+
+		// Return an error if an unsupported authMechanismProperty is specified
+		// for MONGODB-OIDC.
+		for prop := range c.Auth.AuthMechanismProperties {
+			switch prop {
+			case auth.AllowedHostsProp, auth.EnvironmentProp, auth.ResourceProp:
+			default:
+				return fmt.Errorf("auth mechanism property %q is not valid for MONGODB-OIDC", prop)
+			}
+		}
+
 		if env, ok := c.Auth.AuthMechanismProperties[auth.EnvironmentProp]; ok {
 			switch env {
 			case auth.GCPEnvironmentValue, auth.AzureEnvironmentValue:
+				if c.Auth.AuthMechanismProperties[auth.ResourceProp] == "" {
+					return fmt.Errorf("%q must be set for the %s %q", auth.ResourceProp, env, auth.EnvironmentProp)
+				}
+				fallthrough
+			case auth.K8SEnvironmentValue:
 				if c.Auth.OIDCMachineCallback != nil {
 					return fmt.Errorf("OIDCMachineCallback cannot be specified with the %s %q", env, auth.EnvironmentProp)
 				}
 				if c.Auth.OIDCHumanCallback != nil {
 					return fmt.Errorf("OIDCHumanCallback cannot be specified with the %s %q", env, auth.EnvironmentProp)
 				}
-				if c.Auth.AuthMechanismProperties[auth.ResourceProp] == "" {
-					return fmt.Errorf("%q must be set for the %s %q", auth.ResourceProp, env, auth.EnvironmentProp)
-				}
-			default:
+			case auth.TestEnvironmentValue:
 				if c.Auth.AuthMechanismProperties[auth.ResourceProp] != "" {
 					return fmt.Errorf("%q must not be set for the %s %q", auth.ResourceProp, env, auth.EnvironmentProp)
 				}
+				if c.Auth.Username != "" {
+					return fmt.Errorf("must not specify username for %s %q", env, auth.EnvironmentProp)
+				}
+			default:
+				return fmt.Errorf("the %s %q is not supported for MONGODB-OIDC", env, auth.EnvironmentProp)
 			}
 		}
 	}
@@ -665,9 +686,9 @@ func (c *ClientOptions) SetAuth(auth Credential) *ClientOptions {
 
 // SetCompressors sets the compressors that can be used when communicating with a server. Valid values are:
 //
-// 1. "snappy" - requires server version >= 3.4
+// 1. "snappy"
 //
-// 2. "zlib" - requires server version >= 3.6
+// 2. "zlib"
 //
 // 3. "zstd" - requires server version >= 4.2, and driver version >= 1.2.0 with cgo support enabled or driver
 // version >= 1.3.0 without cgo.
@@ -899,9 +920,8 @@ func (c *ClientOptions) SetReplicaSet(s string) *ClientOptions {
 // DeleteManyModel instances to be considered retryable. Unacknowledged writes will not be retried, even if this option
 // is set to true.
 //
-// This option requires server version >= 3.6 and a replica set or sharded cluster and will be ignored for any other
-// cluster type. This can also be set through the "retryWrites" URI option (e.g. "retryWrites=true"). The default is
-// true.
+// This option only works on a replica set or sharded cluster and will be ignored for any other cluster type.
+// This can also be set through the "retryWrites" URI option (e.g. "retryWrites=true"). The default is true.
 func (c *ClientOptions) SetRetryWrites(b bool) *ClientOptions {
 	c.RetryWrites = &b
 
@@ -915,7 +935,7 @@ func (c *ClientOptions) SetRetryWrites(b bool) *ClientOptions {
 // EstimatedDocumentCount, Watch (for Client, Database, and Collection), ListCollections, and ListDatabases. Note that
 // operations run through RunCommand are not retried.
 //
-// This option requires server version >= 3.6 and driver version >= 1.1.0. The default is true.
+// The default is true.
 func (c *ClientOptions) SetRetryReads(b bool) *ClientOptions {
 	c.RetryReads = &b
 
