@@ -33,8 +33,7 @@ func (ioo *InsertOneOptionsBuilder) List() []func(*InsertOneOptions) error {
 
 // SetBypassDocumentValidation sets the value for the BypassDocumentValidation field. If true,
 // writes executed as part of the operation will opt out of document-level validation on the
-// server. This option is valid for MongoDB versions >= 3.2 and is ignored for previous server
-// versions. The default value is false. See https://www.mongodb.com/docs/manual/core/schema-validation/
+// server. The default value is false. See https://www.mongodb.com/docs/manual/core/schema-validation/
 // for more information about document validation.
 func (ioo *InsertOneOptionsBuilder) SetBypassDocumentValidation(b bool) *InsertOneOptionsBuilder {
 	ioo.Opts = append(ioo.Opts, func(opts *InsertOneOptions) error {
@@ -86,8 +85,7 @@ func (imo *InsertManyOptionsBuilder) List() []func(*InsertManyOptions) error {
 
 // SetBypassDocumentValidation sets the value for the BypassDocumentValidation field. If true,
 // writes executed as part of the operation will opt out of document-level validation on the
-// server. This option is valid for MongoDB versions >= 3.2 and is ignored for previous server
-// versions. The default value is false. See https://www.mongodb.com/docs/manual/core/schema-validation/
+// server. The default value is false. See https://www.mongodb.com/docs/manual/core/schema-validation/
 // for more information about document validation.
 func (imo *InsertManyOptionsBuilder) SetBypassDocumentValidation(b bool) *InsertManyOptionsBuilder {
 	imo.Opts = append(imo.Opts, func(opts *InsertManyOptions) error {
