@@ -8,6 +8,7 @@ package mongo
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"strconv"
 
@@ -229,7 +230,8 @@ func (siv SearchIndexView) DropOne(
 		Timeout(siv.coll.client.timeout).Authenticator(siv.coll.client.authenticator)
 
 	err = op.Execute(ctx)
-	if de, ok := err.(driver.Error); ok && de.NamespaceNotFound() {
+	var de driver.Error
+	if errors.As(err, &de) && de.NamespaceNotFound() {
 		return nil
 	}
 	return err
